[Chore/#33] Split egg draw gesture and egg animation into separate files
