+++ DreamEgg/DreamEgg/Stores/EggDrawStore.swift
@@ -1,0 +1,20 @@
+//
+//  EggDrawStore.swift
+//  DreamEgg
+//
+//  Created by Sebin Kwon on 2023/07/21.
+//
+
+import SwiftUI
+
+struct EggDrawStore: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct EggDrawStore_Previews: PreviewProvider {
+    static var previews: some View {
+        EggDrawStore()
+    }
+}
